Set -ansi compiler flag early in configure.ac
The -ansi compiler flag does not enable warnings, but turns of certain
features of GCC. To force feature tests to use this flag, it should be
set early in configure.ac

This is important for the mingw compiler where header files refrain
from declaring certain function (e.g. gettimeofday()) when the -ansi
flag is set.