Check for declaration of gettimeofday()
The function gettimeofday() is not declared in ansi mode with the
mingw compiler, even though the function exists. Therefore, configure must
not only check with AC_CHECK_FUNC, but also with AC_CHECK_DECL.

Also, the checks must run in a C++ context because that is where
gettimeofday() is used.