fix labelToIndex on multiple prefixes
A PDF document can use page label prefixes, for example, to label
only inside a chapter ("A-1" .. "A-<n>", "B-1" .. "B-<m>" et cetera).
When calculating a page index for a given label, poppler would
disregard the page range for all preceding prefixes that do not
match the given prefix.