parseargs: don't use arg->size with GooString argument
40b56994 added GooString arguments but incorrectly used arg->size as
the string length. arg->size is always 0 for GooString arguments. This
worked because the arg->size - 1 string length passed to
GooString::Set() just happened to be the the same value as the
CALC_STRING_LEN default argument indicating the string length is to be
calculated with strlen.
(cherry picked from commit 4e205a6625431dcf95375de009b4354746a4f0b7)